# src/backend_go/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Cognix Golang API Documentation",
        "title": "Cognix API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/auth/google/login": {
            "get": {
                "description": "login using google auth",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "login using google auth",
                "operationId": "auth_login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "redirect base url",
                        "name": "redirect_url",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/chats/create-chat-session": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "creates new chat session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "creates new chat session",
                "operationId": "chat_create_session",
                "parameters": [
                    {
                        "description": "create session parameters",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.CreateChatSession"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChatSession"
                        }
                    }
                }
            }
        },
        "/chats/get-chat-session/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return chat session with messages by given id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "return chat session with messages by given id",
                "operationId": "chat_get_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "session id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChatSession"
                        }
                    }
                }
            }
        },
        "/chats/get-user-chat-sessions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return list of chat sessions for current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "return list of chat sessions for current user",
                "operationId": "chat_get_sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ChatSession"
                            }
                        }
                    }
                }
            }
        },
        "/chats/message/feedback": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add  feedback to message",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "add  feedback to message",
                "operationId": "chat_message_feedback",
                "parameters": [
                    {
                        "description": "send message parameters",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.MessageFeedbackParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChatMessageFeedback"
                        }
                    }
                }
            }
        },
        "/chats/send-message": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "send message and wait stream response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "send message and wait stream response",
                "operationId": "chat_send_message",
                "parameters": [
                    {
                        "description": "send message parameters",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/parameters.CreateChatMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ChatMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ChatMessage": {
            "type": "object",
            "properties": {
                "chat_session_id": {
                    "type": "number"
                },
                "citations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DocumentResponse"
                    }
                },
                "error": {
                    "type": "string"
                },
                "feedback": {
                    "$ref": "#/definitions/model.ChatMessageFeedback"
                },
                "id": {
                    "type": "number"
                },
                "latest_child_message": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_type": {
                    "type": "string"
                },
                "parent_message": {
                    "type": "number"
                },
                "rephrased_query": {
                    "type": "string"
                },
                "time_sent": {
                    "type": "string"
                },
                "token_count": {
                    "type": "integer"
                }
            }
        },
        "model.ChatMessageFeedback": {
            "type": "object",
            "properties": {
                "chat_message_id": {
                    "type": "number"
                },
                "feedback": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "up_votes": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.ChatSession": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChatMessage"
                    }
                },
                "one_shot": {
                    "type": "boolean"
                },
                "persona": {
                    "$ref": "#/definitions/model.Persona"
                },
                "persona_id": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.DocumentResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "link": {
                    "type": "string"
                },
                "message_id": {
                    "type": "number"
                },
                "updated_date": {
                    "type": "string"
                }
            }
        },
...
        "model.LLM": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "endpoint": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "model_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.Persona": {
            "type": "object",
            "properties": {
                "chat_sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ChatSession"
                    }
                },
                "creation_date": {
                    "type": "string"
                },
                "default_persona": {
                    "type": "boolean"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "display_priority": {
                    "type": "integer"
                },
                "id": {
                    "type": "number"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "llm": {
                    "$ref": "#/definitions/model.LLM"
                },
                "llm_id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "prompt": {
                    "$ref": "#/definitions/model.Prompt"
                },
                "starter_messages": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tenant_id": {
                    "type": "string"
                }
            }
        },
        "model.Prompt": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "deleted_date": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "number"
                },
                "last_update": {
                    "$ref": "#/definitions/pg.NullTime"
                },
                "name": {
                    "type": "string"
                },
                "persona_id": {
                    "type": "number"
                },
                "system_prompt": {
                    "type": "string"
                },
                "task_prompt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "parameters.CreateChatMessageRequest": {
            "type": "object",
            "properties": {
                "chat_session_id": {
                    "type": "number"
                },
                "message": {
                    "type": "string"
                },
                "no_ai_answer": {
                    "type": "boolean"
                },
                "parent_message_id": {
                    "type": "number"
                },
                "prompt_id": {
                    "type": "number"
                },
                "query_override": {
                    "description": "RetrievalOptions RetrievalDetails  `json:\"retrieval_options,omitempty\"`",
                    "type": "string"
                },
                "search_doc_ids": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "parameters.CreateChatSession": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "one_shot": {
                    "type": "boolean"
                },
                "persona_id": {
                    "type": "number"
                }
            }
        },
...
        "parameters.MessageFeedbackParam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "vote": {
                    "type": "string"
                }
            }
        },
...
        "pg.NullTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}